third_party/blink/web_tests/platform/webview/external/wpt/pointerevents/pointerevent_after_target_appended_pen-expected.txt

This is a testharness.js-based test.
[FAIL] pointer events from pen received before/after child attached at pointerdown
  assert_equals: events received expected "pointerover@parent,pointerenter@parent,pointerdown@parent,(child-attached),pointerout@parent,pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@parent,pointerenter@parent,pointerdown@parent,(child-attached),pointerover@child,pointerenter@parent,pointerenter@child,pointerout@parent,pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from pen received before/after child attached at pointerup
  assert_equals: events received expected "pointerover@parent,pointerenter@parent,pointerdown@parent,pointerup@parent,(child-attached),pointerout@parent,pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@parent,pointerenter@parent,pointerover@parent,pointerdown@parent,pointerup@parent,(child-attached),pointerout@parent,pointerover@child,pointerenter@child,pointerout@parent,pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from pen received before/after child moved at pointerdown
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,(child-moved),pointerover@child,pointerenter@child,pointerover@child,pointerenter@child,pointerup@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
[FAIL] pointer events from pen received before/after child moved at pointerup
  assert_equals: events received expected "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent" but got "pointerover@child,pointerenter@parent,pointerenter@child,pointerdown@child,pointerup@child,(child-moved),pointerover@child,pointerenter@child,pointerover@child,pointerenter@child,pointerdown@child,pointerup@child,pointerout@child,pointerleave@child,pointerleave@parent"
Harness: the test ran to completion.
